@@ -155,21 +155,21 @@
     "prop-types": "^15.8.1",
     "r-script": "^0.0.4",
     "ramda": "^0.28.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-ace": "^10.1.0",
     "react-avatar-editor": "^13.0.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-google-recaptcha": "^2.1.0",
     "react-grid-layout": "^1.3.4",
-    "react-i18next": "^11.18.1",
+    "react-i18next": "^12.2.2",
     "react-leaflet": "^4.2.1",
     "react-leaflet-markercluster": "^3.0.0-rc1",
     "react-number-format": "^5.1.4",
     "react-redux": "^8.0.2",
-    "react-router": "^6.3.0",
-    "react-router-dom": "^6.3.0",
-    "react-select": "^5.4.0",
+    "react-router": "^6.11.0",
+    "react-router-dom": "^6.11.0",
+    "react-select": "^5.7.3",
     "react-split": "^2.0.14",
     "react-switch": "^7.0.0",
     "react-tiny-popover": "^7.2.4",
